+++ Plugins/Extensions/EpgToXml/__init__.py
@@ -2,7 +2,7 @@
 """EpgToXml Enigma2 plugin package."""
 
 PLUGIN_NAME = "EpgToXml"
-VERSION = "0.5.1"
+VERSION = "0.5.2"
 
 
 def _(text):